rag_engine/requirements.txt
# Core ML/NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
tiktoken>=0.5.2

# LangChain Core
langchain>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1

# LangChain Integrations
langchain-chroma>=0.1.0
langchain-google-genai>=0.0.5
langchain-openai>=0.0.5

# Vector Store
chromadb>=0.4.22

# Web UI
gradio>=5.0.0

# Configuration & Utils
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Data Processing
numpy>=1.24.0
tqdm>=4.66.0

# Testing & Linting
pytest>=7.4.0
pytest-asyncio>=0.21.0
ruff>=0.1.0


